+++ CHANGELOG.md
@@ -4,11 +4,9 @@
 ### Added
 * Support for tags in Databricks pipelines
 ### Fixed
-* n/a
+* Detection of wheel file type in pipeline dependencies when variable is used
 ### Updated
 * `WorkspaceConfigFile` to use source file in user cache instead of base64 content
-### Breaking changes
-* n/a
 
 ## [0.8.5] - 2025-08-04
 ### Added